NullPointerException fix when deleting resource groups

+++ src/main/java/org/dmc/services/ResourceGroupService.java
@@ -55,7 +55,9 @@ public class ResourceGroupService {
 
 		for(String role: roles) {
 			ResourceGroup group = resourceGroupRepository.findByParentTypeAndParentIdAndRole(parentType, parentId, role);
-			resourceGroupRepository.delete(group);
+			if(group != null) {
+				resourceGroupRepository.delete(group);
+			}
 		}
 	}
